breakdownTimeline.php setup with templating
